requirements.txt
pandas
numpy
uvicorn
fastapi
oauth2client
PyOpenSSL
gspread
python-dotenv
pandas